# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "2.0.0",
    "title": "Helpoo Api Documentation",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module, for <b>Helpoo</b>."
  },
  "servers": [
    {
      "url": "http://apidev.helpooapp.net/"
    },
    {
      "url": "https://apidev.helpooapp.net/"
    }
  ],
...
  "paths": {
...
    "/api/v2/drivers/autoChangeDriver": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "driverId": {
                    "example": "any"
                  },
                  "requestId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/drivers/getDriver": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "get driver",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "carServiceTypeId": {
                  "type": "array",
                  "example": [
                    1,
                    2,
                    3
                  ],
                  "items": {
                    "type": "number"
                  }
                },
                "location": {
                  "type": "object",
                  "properties": {
                    "lat": {
                      "type": "number",
                      "example": 31.123
                    },
                    "lng": {
                      "type": "number",
                      "example": 31.123
                    }
                  },
                  "required": [
                    "lat",
                    "lng"
                  ]
                }
              },
              "required": [
                "carServiceTypeId",
                "location"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "carServiceTypeId": {
                    "example": "any"
                  },
                  "location": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/drivers/getDriverOtherServices": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "get driver",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "carServiceTypeId": {
                  "type": "array",
                  "example": [
                    1,
                    2,
                    3
                  ],
                  "items": {
                    "type": "number"
                  }
                },
                "location": {
                  "type": "object",
                  "properties": {
                    "lat": {
                      "type": "number",
                      "example": 31.123
                    },
                    "lng": {
                      "type": "number",
                      "example": 31.123
                    }
                  },
                  "required": [
                    "lat",
                    "lng"
                  ]
                }
              },
              "required": [
                "carServiceTypeId",
                "location"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "carServiceTypeId": {
                    "example": "any"
                  },
                  "location": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/drivers/getDriversService": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "get driver",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "carServiceTypeId": {
                  "type": "array",
                  "example": [
                    1,
                    2,
                    3
                  ],
                  "items": {
                    "type": "number"
                  }
                }
              },
              "required": [
                "carServiceTypeId"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "carServiceTypes": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/drivers/getDriversMap": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/drivers/getOne/{driverId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "driverId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/drivers/assignDriver": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "assign driver",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "requestId": {
                  "type": "number",
                  "example": 1
                },
                "driverId": {
                  "type": "number",
                  "example": 1
                }
              },
              "required": [
                "requestId",
                "driverId"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "requestId": {
                    "example": "any"
                  },
                  "driverId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/drivers/unassignDriver": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "requestId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/drivers/assignVehicle": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assign vehicle",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "driverId": {
                  "type": "number",
                  "example": 1
                },
                "IMEI": {
                  "type": "string",
                  "example": "IMEI"
                },
                "fcmtoken": {
                  "type": "string",
                  "example": "token"
                }
              },
              "required": [
                "driverId",
                "IMEI",
                "fcmtoken"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "driverId": {
                    "example": "any"
                  },
                  "IMEI": {
                    "example": "any"
                  },
                  "fcmtoken": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/drivers/logout": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "logout",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "driverId": {
                  "type": "number",
                  "example": 1
                }
              },
              "required": [
                "driverId"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "driverId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/drivers/driverStats": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/drivers/updateStatus": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "update status",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "driverId": {
                  "type": "number",
                  "example": 1
                },
                "available": {
                  "type": "boolean",
                  "example": true
                }
              },
              "required": [
                "driverId",
                "available"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "driverId": {
                    "example": "any"
                  },
                  "available": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/drivers/getAll": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/drivers/updateDriverPhoto/{id}": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "update driver photo",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "image": {
                  "type": "string",
                  "example": "image"
                }
              },
              "required": [
                "image"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/drivers/updateLocation/{id}": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "update driver location",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "location": {
                  "type": "object",
                  "properties": {
                    "lng": {
                      "type": "number",
                      "example": 1
                    },
                    "lat": {
                      "type": "number",
                      "example": 1
                    }
                  }
                }
              },
              "required": [
                "location"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "location": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/drivers/getDurationAndDistance": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "driverLocation": {
                    "example": "any"
                  },
                  "curClientLocation": {
                    "example": "any"
                  },
                  "prevClientLocation": {
                    "example": "any"
                  },
                  "oldDest": {
                    "example": "any"
                  },
                  "from": {
                    "example": "any"
                  },
                  "reqId": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  },
                  "traffic": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}